backend/requirements.txt
openai-whisper
torch
numpy
sounddevice
soundfile
pvporcupine
fastapi
uvicorn
opencv-python
mss
pytesseract
librosa
scipy
